Bug fix and Refactor extension.js
